Move assertions used in blender and yacat tests to a separate module
