+++ tests/goth/test_run_yacat.py
@@ -2,7 +2,6 @@ import logging
 import math
 import os
 from pathlib import Path
-import re
 
 import pytest
 
@@ -12,6 +11,7 @@ from goth.runner.log import configure_logging
 from goth.runner import Runner
 from goth.runner.probe import RequestorProbe
 
+from .assertions import assert_no_errors, assert_all_invoices_accepted, assert_tasks_processed
 
 logger = logging.getLogger("goth.test.run_yacat")
 
@@ -25,57 +25,14 @@ ALL_TASKS = {"compute_keyspace", "0", f"{CHUNK_SIZE}"}
 # is just a line of output from `yacat.py`.
 
 
-async def assert_no_errors(output_lines: EventStream[str]):
-    """Assert that no output line contains the substring `ERROR`."""
-    async for line in output_lines:
-        if "ERROR" in line:
-            raise AssertionError("Command reported ERROR")
-
-
-async def assert_all_tasks_processed(status: str, output_lines: EventStream[str]):
-    """Assert that for every task in `ALL_TASKS` a line with `Task {status}` will appear."""
-    remaining_tasks = ALL_TASKS.copy()
-
-    async for line in output_lines:
-        m = re.search(rf".*Task {status} .* task data: (.+)$", line)
-        if m:
-            task_data = m.group(1)
-            logger.debug("assert_all_tasks_processed: Task %s: %s", status, task_data)
-            remaining_tasks.discard(task_data)
-        if not remaining_tasks:
-            return
-
-    raise AssertionError(f"Tasks not {status}: {remaining_tasks}")
-
-
 async def assert_all_tasks_started(output_lines: EventStream[str]):
     """Assert that for every task a line with `Task started on provider` will appear."""
-    await assert_all_tasks_processed("started on provider", output_lines)
+    await assert_tasks_processed(ALL_TASKS, "started on provider", output_lines)
 
 
 async def assert_all_tasks_computed(output_lines: EventStream[str]):
     """Assert that for every task a line with `Task computed by provider` will appear."""
-    await assert_all_tasks_processed("finished by provider", output_lines)
-
-
-async def assert_all_invoices_accepted(output_lines: EventStream[str]):
-    """Assert that an invoice is accepted for every provider that confirmed an agreement."""
-    unpaid_agreement_providers = set()
-
-    async for line in output_lines:
-        m = re.search("Agreement confirmed by provider '([^']*)'", line)
-        if m:
-            prov_name = m.group(1)
-            logger.debug("assert_all_invoices_accepted: adding provider '%s'", prov_name)
-            unpaid_agreement_providers.add(prov_name)
-        m = re.search("Accepted invoice from '([^']*)'", line)
-        if m:
-            prov_name = m.group(1)
-            logger.debug("assert_all_invoices_accepted: adding invoice for '%s'", prov_name)
-            unpaid_agreement_providers.remove(prov_name)
-
-    if unpaid_agreement_providers:
-        raise AssertionError(f"Unpaid agreements for: {','.join(unpaid_agreement_providers)}")
+    await assert_tasks_processed(ALL_TASKS, "finished by provider", output_lines)
 
 
 @pytest.mark.asyncio
